chore: remove unused function
